typescript fixes, shc storage refactor
