+++ ui/src/App.tsx
@@ -8,7 +8,6 @@ import QRCode from 'qrcode';
 import QrCodeScannerIcon from '@mui/icons-material/QrCodeScanner';
 import React, { useEffect, useReducer, useRef, useState } from 'react';
 import useDeepCompareEffect from 'use-deep-compare-effect';
-import { useQrDataContext } from './QrDataProvider';
 import * as jose from 'jose';
 
 import {
@@ -240,6 +239,18 @@ function filterForIds(shcIds? : number[]) {
   return (shc: StoredSHC) => shcIds?.includes(shc.id);
 }
 
+export const deriveSHC = (jws : string, id: number) => {
+  const context = new shdc.Context();
+  context.compact = jws;
+  const payload = (shdc.low.decode.jws.compact(context)).jws.payload;
+  const shc: StoredSHC =  {
+    id,
+    jws,
+    payload,
+  };
+  return shc;
+}
+
 let idGenerator = () => Math.random();
 class ServerStateSync {
   dispatch: React.Dispatch<AppAction>;
@@ -399,30 +410,14 @@ export function SHLinkCreate() {
     setIsChecked(updatedCheck);
   }
 
-  const deriveSHC = (jws? : string | null) => {
-    if (jws) {
-      const context = new shdc.Context();
-      context.compact = jws;
-      const payload = (shdc.low.decode.jws.compact(context)).jws.payload;
-      const shc: StoredSHC =  {
-        id: idGenerator(),
-        jws,
-        payload,
-      };
-      return [shc];
-    }
-  }
-
   let oneMonthExpiration = new Date(new Date().getTime() + 1000 * 3600 * 24 * 31);
   let [expiresDate, setExpiresDate] = useState(oneMonthExpiration.toISOString().slice(0, 10));
   let [searchParams] = useSearchParams();
   let custom = Boolean(searchParams.get('custom') === 'true');
-  let scanned = Boolean(searchParams.get('scanned') === 'true');
   let datasetId = Number(searchParams.get('ds'));
   let ds = store.sharing[datasetId];
-  let [datasetName, setDataSetName] = useState(custom ? `Custom Dataset ${Object.keys(store.sharing).length}` : `Scanned Dataset ${Object.keys(store.sharing).length}` );
-  const { jws } = useQrDataContext();
-  let vaccines = (custom ? Object.values(store.vaccines) : scanned ? deriveSHC(jws) : Object.values(store.vaccines).filter(filterForTypes(ds.shcTypes)).filter(filterForIds(ds.shcs)));
+  let [datasetName, setDataSetName] = useState(`Custom Dataset ${Object.keys(store.sharing).length}`);
+  let vaccines = (custom ? Object.values(store.vaccines) : Object.values(store.vaccines).filter(filterForTypes(ds.shcTypes)).filter(filterForIds(ds.shcs)));
   const defaultArray = new Array(vaccines?.length).fill(false)
   // state for keeping track of checked/unchecked vaccines
   const [isChecked, setIsChecked] = useState<boolean[]>(defaultArray);
@@ -439,16 +434,6 @@ export function SHLinkCreate() {
         shlinks: {}
       };
       dispatch({ type: 'dataset-add', ds: customDataSet });
-    } else if (scanned) {
-      datasetId = Object.keys(store.sharing).length;
-      const shcDataSet : DataSet = {
-        id: datasetId,
-        name: datasetName,
-        shcs: vaccines?.map(shc => shc.id),
-        shlinks: {}
-      };
-      vaccines?.forEach((vaccine) => dispatch({ type: 'vaccine-add', vaccine }));
-      dispatch({ type: 'dataset-add', ds: shcDataSet });
     }
     await serverSyncer.createShl(datasetId, {
       passcode: usePasscode ? passcode : undefined,
@@ -461,7 +446,7 @@ export function SHLinkCreate() {
   return (
     <>
       {' '}
-      <h3>New SMART Health Link: {custom ? 'New Custom Dataset' : scanned ? 'New Scanned Dataset' : store.sharing[datasetId].name}</h3>{' '}
+      <h3>New SMART Health Link: {custom ? 'New Custom Dataset' : store.sharing[datasetId].name}</h3>{' '}
       <input
         type="checkbox"
         checked={usePasscode}
@@ -512,9 +497,9 @@ export function SHLinkCreate() {
           }      
         })}
       </ol>
-      {(custom || scanned) && 
+      {(custom) && 
       <>
-        Dataset Name: <input type='text' value={datasetName} onChange={(e) => setDataSetName(e.target.value)} /> <br></br>
+        Custom Dataset Name: <input type='text' value={datasetName} onChange={(e) => setDataSetName(e.target.value)} /> <br></br>
       </>}
       <button onClick={activate}>Activate new sharing link</button>
     </>
